+++ activestorage/lib/active_storage/errors.rb
@@ -23,30 +23,4 @@ module ActiveStorage
   # Raised when ActiveStorage::Blob#download is called on a blob where the
   # backing file is no longer present in its service.
   class FileNotFoundError < Error; end
-
-  if defined?(Rails)
-    require "active_support/actionable_error"
-    require "rails/command"
-
-    # Raised when we detect that Active Storage has not been initialized.
-    class InstallError < Error
-      include ActiveSupport::ActionableError
-
-      def initialize(message = nil)
-        super(message || <<~MESSAGE)
-          Action Storage does not appear to be installed. Do you want to
-          install it now?
-        MESSAGE
-      end
-
-      trigger on: ActiveRecord::StatementInvalid, if: -> error do
-        [Blob, Attachment].any? { |model| error.message.match?(model.table_name) }
-      end
-
-      action "Install now" do
-        Rails::Command.invoke("active_storage:install")
-        Rails::Command.invoke("db:migrate")
-      end
-    end
-  end
 end